fix: co-locate types
